CDC weight-for-age (https://www.cdc.gov/growthcharts/percentile_data_files.htm) Weight-for-age charts, 0 to 20 years, LMS parameters and selected smoothed weight percentiles in kilograms, by sex and age Stature-for-age charts, 0 to 20 years, LMS parameters and selected smoothed weight percentiles in cm, by sex and age Insert citation notes here in the future

CDC weight-for-age (https://www.cdc.gov/growthcharts/percentile_data_files.htm) Weight-for-age charts, 0 to 20 years, LMS parameters and selected smoothed weight percentiles in kilograms, by sex and age Stature-for-age charts, 0 to 20 years, LMS parameters and selected smoothed weight percentiles in cm, by sex and age Insert citation notes here in the future

Usage

cdc.expand

Format

cdc.expand

data.frame': 97177 obs. of 6 variables: $ SEX : num 0 0 0 0 0 0 0 0 0 0 ... $ AGEMO: num 0 0.5 1.5 2.5 3.5 4.5 5.5 6.5 7.5 8.5 ... $ AGEYR: num 0 0.0417 0.125 0.2083 0.2917 ... $ WT : num 2.6 3.8 4.9 4.9 6.5 7.6 7.2 8.4 8.8 7.2 ... $ HT : num 45.7 51.8 56.6 57 62.5 65.9 64.8 68.5 69.9 65.8 ... $ BMI : num 12.4 14.2 15.3 15.1 16.6 ...

Source

https://www.cdc.gov/growthcharts/percentile_data_files.htm